+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Added]
 ### [Changed]
+- webpack-packages/templating-plugin - Changed str.replace algorithm when prerendering to deal with large strings.
+
 ### [Deprecated]
 ### [Removed]
 ### [Fixed]